+++ tests/commonjs.js
@@ -104,12 +104,7 @@ for (const _PATH of ['../packages/core-js-pure', '../packages/core-js']) {
   ok('next' in load('features/array/virtual/iterator').call([]));
   ok(load('features/array/virtual').includes.call([1, 2, 3], 2));
   ok('from' in load('features/array'));
-  ok(load('features/array/pop')([1, 2, 3]) === 3);
-  ok(load('features/array/push')([1, 2, 3], 4) === 4);
-  ok(load('features/array/reverse')([1, 2, 3])[0] === 3);
-  ok(load('features/array/shift')([1, 2, 3]) === 1);
   ok(load('features/array/splice')([1, 2, 3], 1, 2)[0] === 2);
-  ok(load('features/array/unshift')([1, 2, 3], 0) === 4);
   ok(load('features/math/acosh')(1) === 0);
   ok(Object.is(load('features/math/asinh')(-0), -0));
   ok(load('features/math/atanh')(1) === Infinity);